garden-server/src/requirements.txt
flask
flask_cors
smbus
smbus2
RPi.GPIO
Flask-JWT-Extended
picamera
opencv-python
Pillow
mysql-connector-python